backend/requirements.txt
astroid==2.2.5
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
ecdsa==0.13.2
Flask==2.2.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==3.0.2
future==0.17.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
isort==4.3.18
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.4.0
MarkupSafe==2.1.2
mccabe==0.6.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.3.1
pylint==2.3.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
requests==2.28.2
rsa==4.9
six==1.12.0
SQLAlchemy==2.0.3
typed-ast==1.4.2
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.11.1
zipp==3.13.0
